Split out RoutineContext into it's own file
