+++ RoutineManager.cs
@@ -25,8 +25,6 @@ using UnityEngine;
 
 namespace Routines
 {
-	public delegate void OnProgressDelegate(float progress);
-
 	public class RoutineManager : MonoBehaviour, IRoutineContext
 	{
 		private RoutineContext context = new RoutineContext();
@@ -79,135 +77,6 @@ namespace Routines
 		public IEnumerator WaitForCustomYieldInstruction(CustomYieldInstruction yieldInstruction)
 		{
 			return context.WaitForCustomYieldInstruction(yieldInstruction);
-		}
-	}
-
-	public class RoutineContext : IRoutineContext
-	{
-		private class NextFrameResumable : Resumable
-		{
-			public List<IResumer> newResumers = new List<IResumer>();
-			public List<IResumer> resumers = new List<IResumer>();
-
-			public override void Run(IResumer resumer)
-			{
-				newResumers.Add(resumer);
-			}
-
-			public void Resume()
-			{
-				foreach (var resumer in resumers)
-				{
-					resumer.Resume();
-				}
-				resumers.Clear();
-			}
-
-			public void Flush()
-			{
-				resumers.AddRange(newResumers);
-				newResumers.Clear();
-			}
-		}
-
-		private NextFrameResumable nextFrameResumable = new NextFrameResumable();
-		private List<Routine> activeRoutines = new List<Routine>();
-
-		public void Update()
-		{
-			for (var i = 0; i < activeRoutines.Count;)
-			{
-				var routine = activeRoutines[i];
-				if (routine.IsDone)
-				{
-					Routine.Release(ref routine);
-					activeRoutines.RemoveAt(i);
-				}
-				else
-				{
-					++i;
-				}
-			}
-
-			nextFrameResumable.Resume();
-		}
-
-		public void LateUpdate()
-		{
-			nextFrameResumable.Flush();
-		}
-
-		public void OnDestroy()
-		{
-			StopAllRoutines();
-		}
-
-		public Routine.Handle RunRoutine(IEnumerator enumerator, System.Action onStop = null)
-		{
-			return RunRoutine(enumerator, null);
-		}
-
-		public Routine.Handle RunRoutine(IEnumerator enumerator, Object context, System.Action onStop = null)
-		{
-			var routine = Routine.Create();
-			routine.Start(enumerator, context, null, onStop);
-			activeRoutines.Add(routine);
-			return routine.GetHandle();
-		}
-
-		public void StopAllRoutines()
-		{
-			for (var i = 0; i < activeRoutines.Count; ++i)
-			{
-				var routine = activeRoutines[i];
-				Routine.Release(ref routine);
-			}
-			activeRoutines.Clear();
-		}
-
-		public IEnumerator WaitForNextFrame()
-		{
-			return nextFrameResumable;
-		}
-
-		public IEnumerator WaitForSeconds(float seconds)
-		{
-			while (seconds > 0.0f)
-			{
-				yield return nextFrameResumable;
-				seconds -= Time.deltaTime;
-			}
-		}
-
-		public IEnumerator WaitUntilCondition(System.Func<bool> condition)
-		{
-			while (!condition())
-			{
-				yield return nextFrameResumable;
-			}
-		}
-
-		public IEnumerator WaitForAsyncOperation(AsyncOperation asyncOperation, OnProgressDelegate onProgress = null)
-		{
-			var lastProgress = (onProgress != null) ? asyncOperation.progress : float.MaxValue;
-			while (!asyncOperation.isDone)
-			{
-				yield return nextFrameResumable;
-
-				if (asyncOperation.progress > lastProgress)
-				{
-					lastProgress = asyncOperation.progress;
-					onProgress(lastProgress);
-				}
-			}
-		}
-
-		public IEnumerator WaitForCustomYieldInstruction(CustomYieldInstruction yieldInstruction)
-		{
-			while (yieldInstruction.keepWaiting)
-			{
-				yield return nextFrameResumable;
-			}
 		}
 	}
 }
